% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_psi_pop_x}
\alias{mean_psi_pop_x}
\title{Sample expectation of PSPA psi-X}
\usage{
mean_psi_pop_x(
  X_lab,
  Xhat_lab,
  Xhat_unlab,
  Y_lab,
  Y_unlab,
  w,
  theta,
  quant = NA,
  A_lab,
  Ahat_lab,
  Ahat_unlab,
  method
)
}
\arguments{
\item{X_lab}{Array or DataFrame containing observed covariates in labeled data.}

\item{Xhat_lab}{Array or DataFrame containing predicted covariates in labeled data.}

\item{Xhat_unlab}{Array or DataFrame containing predicted covariates in unlabeled data.}

\item{Y_lab}{Array or DataFrame of observed outcomes in labeled data.}

\item{Y_unlab}{Array or DataFrame of observed outcomes in unlabeled data.}

\item{w}{weights vector PSPA linear regression (d-dimensional, where d equals the number of covariates).}

\item{theta}{parameter theta}

\item{quant}{quantile for quantile estimation}

\item{method}{indicates the method to be used for M-estimation. Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
sample expectation of PSPA psi
}
\description{
\code{mean_psi_pop_x} function for sample expectation of PSPA psi
}
