% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optim_weights_x}
\alias{optim_weights_x}
\title{One-step update for obtaining the weight vector}
\usage{
optim_weights_x(
  X_lab,
  Xhat_lab,
  Xhat_unlab,
  Y_lab,
  Y_unlab,
  w,
  theta,
  quant = NA,
  method
)
}
\arguments{
\item{X_lab}{Array or DataFrame containing observed covariates in labeled data.}

\item{Xhat_lab}{Array or DataFrame containing predicted covariates in labeled data.}

\item{Xhat_unlab}{Array or DataFrame containing predicted covariates in unlabeled data.}

\item{Y_lab}{Array or DataFrame of observed outcomes in labeled data.}

\item{Y_unlab}{Array or DataFrame of observed outcomes in unlabeled data.}

\item{w}{weights vector PSPA linear regression (d-dimensional, where d equals the number of covariates).}

\item{theta}{parameter theta}

\item{quant}{quantile for quantile estimation}

\item{method}{indicates the method to be used for M-estimation. Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
weights
}
\description{
\code{optim_weights} function for One-step update for obtaining estimator
}
