% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pspa_x}
\alias{pspa_x}
\title{PSPA M-Estimation for ML-predicted covariates}
\usage{
pspa_x(
  X_lab,
  Xhat_lab,
  Xhat_unlab,
  Y_lab,
  Y_unlab,
  alpha = 0.05,
  weights = NA,
  quant = NA,
  intercept = FALSE,
  method
)
}
\arguments{
\item{X_lab}{Array or DataFrame containing observed covariates in labeled data.}

\item{Xhat_lab}{Array or DataFrame containing predicted covariates in labeled data.}

\item{Xhat_unlab}{Array or DataFrame containing predicted covariates in unlabeled data.}

\item{Y_lab}{Array or DataFrame of observed outcomes in labeled data.}

\item{Y_unlab}{Array or DataFrame of observed outcomes in unlabeled data.}

\item{alpha}{Specifies the confidence level as 1 - alpha for confidence intervals.}

\item{weights}{weights vector PSPA linear regression (d-dimensional, where d equals the number of covariates).}

\item{quant}{quantile for quantile estimation}

\item{intercept}{Boolean indicating if the input covariates' data contains the intercept (TRUE if the input data contains)}

\item{method}{indicates the method to be used for M-estimation. Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
A summary table presenting point estimates, standard error, confidence intervals (1 - alpha), P-values, and weights.
}
\description{
\code{pspa_x} function conducts post-prediction M-Estimation.
}
\examples{
data <- sim_data_x()
X_lab <- data$X_lab
Xhat_lab <- data$Xhat_lab
Xhat_unlab <- data$Xhat_unlab
Y_lab <- data$Y_lab
Y_unlab <- data$Y_unlab
pspa_x(X_lab = X_lab, Xhat_lab = Xhat_lab,
      Xhat_unlab = Xhat_unlab, Y_lab = Y_lab, Y_unlab = Y_unlab,
      alpha = 0.05, method = "ols")
}
