% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Sigma_cal_x}
\alias{Sigma_cal_x}
\title{Variance-covariance matrix of the estimation equation-X}
\usage{
Sigma_cal_x(
  X_lab,
  Xhat_lab,
  Xhat_unlab,
  Y_lab,
  Y_unlab,
  w,
  theta,
  quant = NA,
  A_lab,
  Ahat_lab,
  Ahat_unlab,
  method
)
}
\arguments{
\item{X_lab}{Array or DataFrame containing observed covariates in labeled data.}

\item{Xhat_lab}{Array or DataFrame containing predicted covariates in labeled data.}

\item{Xhat_unlab}{Array or DataFrame containing predicted covariates in unlabeled data.}

\item{Y_lab}{Array or DataFrame of observed outcomes in labeled data.}

\item{Y_unlab}{Array or DataFrame of observed outcomes in unlabeled data.}

\item{w}{weights vector PSPA linear regression (d-dimensional, where d equals the number of covariates).}

\item{theta}{parameter theta}

\item{quant}{quantile for quantile estimation}

\item{A_lab}{A using labeled data}

\item{Ahat_lab}{Ahat using labeled data}

\item{Ahat_unlab}{Ahat using unlabeled data}

\item{method}{indicates the method to be used for M-estimation. Options include "mean", "quantile", "ols", "logistic", and "poisson".}
}
\value{
variance-covariance matrix of the estimation equation
}
\description{
\code{Sigma_cal_x} function for variance-covariance matrix of the estimation equation
}
