% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_data_x}
\alias{sim_data_x}
\title{Simulate the data for testing the functions}
\usage{
sim_data_x(r = 0.9, binary = FALSE, n_factor = 1)
}
\arguments{
\item{r}{imputation correlation}

\item{binary}{simulate binary outcome or not}
}
\value{
simulated data
}
\description{
\code{sim_data_x} for simulation with ML-predicted X
}
